Update Avatax migration scripts
