+++ apps/avatax/README.md
@@ -86,11 +86,11 @@ You need to set `REST_APL_TOKEN` & `REST_APL_ENDPOINT` in our `.env` file first.
 Test migration with dry run, operation will not modify any data:
 
 ```bash
-pnpm dlx tsx ./scripts/migrations/run-webhooks-migration-dry-run.ts
+pnpm migrate:dry-run
 ```
 
 To start the migration run command:
 
 ```bash
-pnpm dlx tsx scripts/migrations/run-webhooks-migration.ts
+pnpm migrate
 ```
