+++ apps/avatax/scripts/deploy.ts
@@ -1,0 +1,3 @@
+import { execSync } from "child_process";
+
+execSync("pnpm run migrate", { stdio: "inherit" });
