+++ apps/avatax/src/modules/avatax/calculate-taxes/avatax-calculate-taxes-adapter.ts
@@ -1,14 +1,14 @@
 import { AuthData } from "@saleor/app-sdk/APL";
+import { createLogger } from "../../../logger";
 import { ClientLogger } from "../../logs/client-logger";
 import { CalculateTaxesResponse } from "../../taxes/tax-provider-webhook";
 import { WebhookAdapter } from "../../taxes/tax-webhook-adapter";
+import { CalculateTaxesPayload } from "../../webhooks/payloads/calculate-taxes-payload";
 import { AvataxClient, CreateTransactionArgs } from "../avatax-client";
 import { AvataxConfig } from "../avatax-connection-schema";
 import { normalizeAvaTaxError } from "../avatax-error-normalizer";
 import { AvataxCalculateTaxesPayloadService } from "./avatax-calculate-taxes-payload.service";
 import { AvataxCalculateTaxesResponseTransformer } from "./avatax-calculate-taxes-response-transformer";
-import { createLogger } from "../../../logger";
-import { CalculateTaxesPayload } from "../../webhooks/calculate-taxes-payload";
 
 export type AvataxCalculateTaxesTarget = CreateTransactionArgs;
 export type AvataxCalculateTaxesResponse = CalculateTaxesResponse;
