+++ apps/avatax/src/modules/avatax/calculate-taxes/avatax-calculate-taxes-payload-transformer.ts
@@ -1,5 +1,6 @@
 import { DocumentType } from "avatax/lib/enums/DocumentType";
 import { discountUtils } from "../../taxes/discount-utils";
+import { CalculateTaxesPayload } from "../../webhooks/payloads/calculate-taxes-payload";
 import { avataxAddressFactory } from "../address-factory";
 import { AvataxClient, CreateTransactionArgs } from "../avatax-client";
 import { AvataxConfig, defaultAvataxConfig } from "../avatax-connection-schema";
@@ -7,7 +8,6 @@ import { avataxCustomerCode } from "../avatax-customer-code-resolver";
 import { AvataxEntityTypeMatcher } from "../avatax-entity-type-matcher";
 import { AvataxTaxCodeMatches } from "../tax-code/avatax-tax-code-match-repository";
 import { AvataxCalculateTaxesPayloadLinesTransformer } from "./avatax-calculate-taxes-payload-lines-transformer";
-import { CalculateTaxesPayload } from "../../webhooks/calculate-taxes-payload";
 
 export class AvataxCalculateTaxesPayloadTransformer {
   private matchDocumentType(config: AvataxConfig): DocumentType {
