+++ apps/avatax/src/modules/avatax/calculate-taxes/avatax-calculate-taxes-payload.service.ts
@@ -1,9 +1,9 @@
 import { AuthData } from "@saleor/app-sdk/APL";
+import { CalculateTaxesPayload } from "../../webhooks/payloads/calculate-taxes-payload";
 import { CreateTransactionArgs } from "../avatax-client";
 import { AvataxConfig } from "../avatax-connection-schema";
 import { AvataxTaxCodeMatchesService } from "../tax-code/avatax-tax-code-matches.service";
 import { AvataxCalculateTaxesPayloadTransformer } from "./avatax-calculate-taxes-payload-transformer";
-import { CalculateTaxesPayload } from "../../webhooks/calculate-taxes-payload";
 
 export class AvataxCalculateTaxesPayloadService {
   constructor(private authData: AuthData) {}
