+++ apps/avatax/src/modules/webhooks/definitions/checkout-calculate-taxes.ts
@@ -1,0 +1,12 @@
+import { SaleorSyncWebhook } from "@saleor/app-sdk/handlers/next";
+import { UntypedCalculateTaxesDocument } from "../../../../generated/graphql";
+import { saleorApp } from "../../../../saleor-app";
+import { CalculateTaxesPayload } from "../payloads/calculate-taxes-payload";
+
+export const checkoutCalculateTaxesSyncWebhook = new SaleorSyncWebhook<CalculateTaxesPayload>({
+  name: "CheckoutCalculateTaxes",
+  apl: saleorApp.apl,
+  event: "CHECKOUT_CALCULATE_TAXES",
+  query: UntypedCalculateTaxesDocument,
+  webhookPath: "/api/webhooks/checkout-calculate-taxes",
+});
