+++ apps/avatax/src/modules/webhooks/payloads/calculate-taxes-payload.ts
@@ -1,4 +1,4 @@
-import { CalculateTaxesEventFragment } from "../../../generated/graphql";
+import { CalculateTaxesEventFragment } from "../../../../generated/graphql";
 
 /**
  * Common subset for ORDER and CHECKOUT events
