+++ apps/avatax/src/modules/webhooks/payloads/order-cancelled-payload.ts
@@ -1,0 +1,6 @@
+import { OrderCancelledEventSubscriptionFragment } from "../../../../generated/graphql";
+
+export type OrderCancelledPayload = Extract<
+  OrderCancelledEventSubscriptionFragment,
+  { __typename: "OrderCancelled" }
+>;
