+++ apps/avatax/src/modules/webhooks/payloads/order-confirmed-payload.ts
@@ -1,0 +1,6 @@
+import { OrderConfirmedEventSubscriptionFragment } from "../../../../generated/graphql";
+
+export type OrderConfirmedPayload = Extract<
+  OrderConfirmedEventSubscriptionFragment,
+  { __typename: "OrderConfirmed" }
+>;
