+++ apps/avatax/src/pages/api/webhooks/checkout-calculate-taxes.ts
@@ -1,31 +1,20 @@
-import { SaleorSyncWebhook } from "@saleor/app-sdk/handlers/next";
-import { UntypedCalculateTaxesDocument } from "../../../../generated/graphql";
-import { saleorApp } from "../../../../saleor-app";
+import { withOtel } from "@saleor/apps-otel";
+import * as Sentry from "@sentry/nextjs";
+import { createLogger } from "../../../logger";
 import { WebhookResponse } from "../../../modules/app/webhook-response";
 import { getActiveConnectionService } from "../../../modules/taxes/get-active-connection-service";
-import { withOtel } from "@saleor/apps-otel";
-import { createLogger } from "../../../logger";
-import * as Sentry from "@sentry/nextjs";
 import { calculateTaxesErrorsStrategy } from "../../../modules/webhooks/calculate-taxes-errors-strategy";
 
-import { verifyCalculateTaxesPayload } from "../../../modules/webhooks/validate-webhook-payload";
-import { CalculateTaxesPayload } from "../../../modules/webhooks/calculate-taxes-payload";
 import { wrapWithLoggerContext } from "@saleor/apps-logger/node";
 import { loggerContext } from "../../../logger-context";
+import { checkoutCalculateTaxesSyncWebhook } from "../../../modules/webhooks/definitions/checkout-calculate-taxes";
+import { verifyCalculateTaxesPayload } from "../../../modules/webhooks/validate-webhook-payload";
 
 export const config = {
   api: {
     bodyParser: false,
   },
 };
-
-export const checkoutCalculateTaxesSyncWebhook = new SaleorSyncWebhook<CalculateTaxesPayload>({
-  name: "CheckoutCalculateTaxes",
-  apl: saleorApp.apl,
-  event: "CHECKOUT_CALCULATE_TAXES",
-  query: UntypedCalculateTaxesDocument,
-  webhookPath: "/api/webhooks/checkout-calculate-taxes",
-});
 
 /**
  * TODO: Add tests to handler
