+++ apps/avatax/src/pages/api/webhooks/order-calculate-taxes.ts
@@ -1,30 +1,19 @@
-import { SaleorSyncWebhook } from "@saleor/app-sdk/handlers/next";
-import { UntypedCalculateTaxesDocument } from "../../../../generated/graphql";
-import { saleorApp } from "../../../../saleor-app";
-import { getActiveConnectionService } from "../../../modules/taxes/get-active-connection-service";
-import { WebhookResponse } from "../../../modules/app/webhook-response";
-import { withOtel } from "@saleor/apps-otel";
-import { createLogger } from "../../../logger";
-import { calculateTaxesErrorsStrategy } from "../../../modules/webhooks/calculate-taxes-errors-strategy";
-import * as Sentry from "@sentry/nextjs";
-import { verifyCalculateTaxesPayload } from "../../../modules/webhooks/validate-webhook-payload";
-import { CalculateTaxesPayload } from "../../../modules/webhooks/calculate-taxes-payload";
 import { wrapWithLoggerContext } from "@saleor/apps-logger/node";
+import { withOtel } from "@saleor/apps-otel";
+import * as Sentry from "@sentry/nextjs";
+import { createLogger } from "../../../logger";
 import { loggerContext } from "../../../logger-context";
+import { WebhookResponse } from "../../../modules/app/webhook-response";
+import { getActiveConnectionService } from "../../../modules/taxes/get-active-connection-service";
+import { calculateTaxesErrorsStrategy } from "../../../modules/webhooks/calculate-taxes-errors-strategy";
+import { orderCalculateTaxesSyncWebhook } from "../../../modules/webhooks/definitions/order-calculate-taxes";
+import { verifyCalculateTaxesPayload } from "../../../modules/webhooks/validate-webhook-payload";
 
 export const config = {
   api: {
     bodyParser: false,
   },
 };
-
-export const orderCalculateTaxesSyncWebhook = new SaleorSyncWebhook<CalculateTaxesPayload>({
-  name: "OrderCalculateTaxes",
-  apl: saleorApp.apl,
-  event: "ORDER_CALCULATE_TAXES",
-  query: UntypedCalculateTaxesDocument,
-  webhookPath: "/api/webhooks/order-calculate-taxes",
-});
 
 export default wrapWithLoggerContext(
   withOtel(
